requirements.txt
openai
fastapi
uvicorn
fireworks-ai
pydantic
python-dotenv
requests
fireworks
gtts




